requirements.txt
boto3==1.6.3
botocore==1.9.3
Brlapi==0.6.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
cryptography==1.7.1
cupshelpers==1.0
dash==0.21.0
dash-core-components==0.20.1
dash-html-components==0.9.0
dash-renderer==0.11.3
decorator==4.2.1
docutils==0.14
Flask==0.12.2
Flask-Caching==1.3.3
Flask-Compress==1.4.0
gunicorn==19.7.1
httplib2==0.9.2
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
keyring==10.1
keyrings.alt==1.3
louis==3.0.0
lxml==4.1.1
Mako==1.0.6
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.1
paho-mqtt==1.3.1
pandas==0.22.0
pandas-datareader==0.6.0
pexpect==4.2.1
Pillow==4.0.0
plotly==2.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.1
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.22.0
pysmbc==1.0.15.6
python-apt==1.4.0b3
python-dateutil==2.6.1
python-debian==0.1.30
python-debianbts==2.6.1
pytz==2018.3
pyxdg==0.25
redis==2.10.6
reportbug==7.1.7
reportlab==3.3.0
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
s3transfer==0.1.13
SecretStorage==2.3.1
six==1.11.0
traitlets==4.3.2
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
wrapt==1.10.11
